THarris README.md
Creative, detail-oriented, software engineer with a deep interest in web3. Continuous learning in order to create and implement successful front and back end web applications. Looking to bring my skills to a tech company with global reach.

Coming from a competitive background, the importance of performance is of high priority. The road ahead may not be the easiest, but that is all the more reason to get on the good foot.